requirements.txt
openai==1.13.3
httpx==0.27.0
httpcore==1.0.4
pydantic>=2.0
python-dotenv
tqdm
anyio
matplotlib
pandas
fastapi
uvicorn
